Back-End fixed

+++ PetShop.Core/DomainService/IRepository.cs
@@ -8,9 +8,9 @@ namespace CompanyName.PetShop.RestApi.Data
     public interface IRepository<T>
     {
         IEnumerable<T> GetAll();
-        T Get(long id);
+        T Get(int id);
         void Add(T entity);
         void Edit(T entity);
-        void Remove(long id);
+        void Remove(int id);
     }
 }
